requirements.txt
Bio==1.8.0
biopython==1.81
fair_esm==2.0.0
matplotlib==3.5.3
modlamp==4.3.0
numpy==1.19.5
pandas==1.3.5
scikit_learn==1.0.2
scipy==1.4.1
seaborn==0.13.2
tensorflow==2.2.0
torch==1.13.1
tqdm==4.66.4
shifterator==0.3.0
git+https://gitlab.com/vail-uvm/vailtools.git@develop

